Replace original scheduling workflow
