+++ platform/api-server/api/controllers/platform-api-controller.js
@@ -3,7 +3,6 @@ const fs = require('fs');
 const fsExtra = require('fs-extra');
 const path = require('path');
 const utils = require('../../../utils/utils');
-const child_process = require('child_process');
 const MqttController = require('../../../utils/mqtt-controller');
 const mqttController = MqttController.getInstance();
 const mqttTopic = 'platform-data';
@@ -115,82 +114,18 @@ exports.deployApplication = async function(req, res) {
     console.log(`Received deploy metadata at ${deployMetadataPath}`);
     console.log(`Received app package at ${packagePath}`);
 
-    // Unpackage the app metadata.
-    const extractDir = '/tmp';
-    try {
-        child_process.execFileSync(
-            utils.tarPath,
-            ['-x', '-f', packagePath, '_metadata.json'],
-            { cwd: extractDir, timeout: 1000 });
-    } catch (e) {
-        console.log(`Failed to unpackage app metadata: ${e}.`);
-        res.sendStatus(500);
-        return;
-    }
-
-    const runMetadata = JSON.parse(await fs.promises.readFile(path.join(extractDir, '_metadata.json')));
-    const deployMetadata = JSON.parse(await fs.promises.readFile(deployMetadataPath));
-
-    // Obtain gateway resource information to make a scheduling decision.
-    const graph = await utils.getLinkGraph();
-    const gatewayResources = await Promise.all(Object.keys(graph.data).map((gatewayId) => {
-        const gatewayInfo = graph.data[gatewayId];
-        const ip = gatewayInfo.ip;
-
-        // Get a count of devices by type for each gateway.
-        // This is used for deployment metadata that specifies a particular kind of device.
-        var deviceTypes = new Map();
-        gatewayInfo.devices.forEach((device) => {
-            if (deviceTypes.has(device.type)) {
-                deviceTypes[device.type] += 1;
-            } else {
-                deviceTypes.set(device.type, 1);
-            }
+    const executeResult = await messagingService.query(
+        serviceName, 'app-manager', 'execute-app', {
+            'packagePath': packagePath,
+            'deployMetadataPath': deployMetadataPath
         });
 
-        const opts = {
-            method: 'GET',
-            uri: `http://${ip}:5000/gateway/resources`,
-            json: true
-        };
-
-        return request(opts)
-            .then((resources) => {
-                const deviceInfo = {
-                    id: gatewayId,
-                    ip: ip,
-                    resources: resources,
-                    deviceIds: gatewayInfo.devices.map(device => device.id),
-                    deviceTypes: deviceTypes
-                };
-
-                return deviceInfo;
-            });
-    }));
-
-    // Run the scheduling algorithm to determine where to put the application.
-    const gateway = schedule(deployMetadata, runMetadata, gatewayResources);
-    if (gateway !== null) {
-        console.log(`Sending application to run on ${gateway.ip}.`);
-
-        const gatewayUri = `http://${gateway.ip}:5000/gateway/execute-app-v2`;
-        const formData = {
-            'appPackage': fsExtra.createReadStream(packagePath),
-            'deployMetadata': fsExtra.createReadStream(deployMetadataPath)
-        };
-        const opts = {
-            method: 'POST',
-            uri: gatewayUri,
-            formData: formData
-        };
-
-        request(opts)
-            .then(() => { res.sendStatus(204); },
-                  () => { res.sendStatus(500); });
+    if (executeResult.status === true) {
+        res.status(204);
+    } else if (executeResult.message.length !== 0) {
+        res.status(400).send(executeResult.message);
     } else {
-        // No gateways available for the application.
-        // Send back 503 Service Unavailable.
-        res.sendStatus(503);
+        res.status(500);
     }
 };
